Revert "RDKAAMP-3783  Expand test coverage for files under plugins/"
This reverts commit 1ce6bdce0355b8e41e78d8c83363c4149280f42c.

Reason for revert: temporarily reverting due to SEGFAULT

+++ plugins/gst-plugins-rdk-aamp/drm/gst/gstaampcdmidecryptor.cpp
@@ -903,7 +903,7 @@ static GstFlowReturn gst_aampcdmidecryptor_transform_ip(
 	        goto free_resources;
 	    }
 
-	    GST_TRACE_OBJECT(aampcdmidecryptor, "position: %d, size: %lu", position,
+	    GST_TRACE_OBJECT(aampcdmidecryptor, "position: %d, size: %d", position,
 	            map.size);
 
 	    // collect all the encrypted bytes into one contiguous buffer
@@ -1163,7 +1163,7 @@ static gboolean gst_aampcdmidecryptor_sink_event(GstBaseTransform * trans,
         const gchar* systemId;
         const gchar* origin;
         unsigned char *outData = NULL;
-        size_t outDataLen = 0;
+        unsigned int outDataLen = 0;
         GstBuffer* initdatabuffer;
 
 	if(NULL == aampcdmidecryptor)
